Renamed PlatformModule
